requirements.txt
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
loguru==0.7.0
redis==4.5.4
wordfreq==3.0.3
debugpy==1.6.7
python-telegram-bot-calendar==1.0.5
